Fix horrifying deadlock in cond_wait
Before the code called mutex_lock while holding the cond's spinlock.  This is
bad, as mutex_lock can sleep for arbitrary amounts of time (and certainly more
than O(1) time).

But, the corresponding call to mutex_unlock above is safe and necessary.  It
is safe because mutex_unlock is O(1) and never sleeps.  It is necessary
because of pthread_cond_wait's atomicity guarantee: that another thread
getting the mutex lock after us and performing an action on the cond happens
as if we have already finished our cond operation.